+++ controllers/user.controller.js
@@ -4,12 +4,32 @@ const UsersDAO = require("../dao/usersDAO")
 
 const hashPassword = async password => await bcrypt.hash(password, 10)
 
+const verify = async(userJwt) => {
+    return jwt.verify(userJwt, process.env.SECRET_KEY, (error, res) => {
+        if (error) {
+            return false
+        }
+        return true
+    })
+}
+
+const validateToken = async (bearer) => {
+    if(bearer !== undefined ) {
+        const jwt = bearer.split(' ')[1]
+        return await verify(jwt)
+    }
+    else {
+        return false
+    }
+}
+
 class User {
-    constructor({ username, password, name, surname = {} } = {}) {
+    constructor({ username, password, name, surname, personal = {} } = {}) {
         this.username = username
         this.password = password
         this.name = name
         this.surname = surname
+        this.personal = personal
     }
 
     async comparePassword(plainText) {
@@ -17,7 +37,7 @@ class User {
     }
 
     toJson() {
-        return { name: this.name, surname: this.surname, username: this.username }
+        return { name: this.name, surname: this.surname, username: this.username, personal: this.personal }
     }
 
     encoded() {
@@ -29,30 +49,12 @@ class User {
             process.env.SECRET_KEY,
         )
     }
-
-    static async decoded(userJwt) {
-        return jwt.verify(userJwt, process.env.SECRET_KEY, (error, res) => {
-            if (error) {
-                return { error }
-            }
-            return new User(res)
-        })
-    }
 }
 
 class UsersController {
     static async login(req, res) {
         try {
             const { username, password } = req.body
-
-            if (!username || typeof username !== "string") {
-                res.status(400).json({ error: "Bad username format, expected string." })
-                return
-            }
-            if (!password || typeof password !== "string") {
-                res.status(400).json({ error: "Bad password format, expected string." })
-                return
-            }
 
             let userData = await UsersDAO.getUser(username)
 
@@ -68,56 +70,9 @@ class UsersController {
                 return
             }
 
-            const loginResponse = await UsersDAO.loginUser(user.username, user.encoded())
-            if (!loginResponse.success) {
-                res.status(500).json({ error: loginResponse.error })
-                return
-            }
             // On success return the auth_token and all the user info
-            res.json({ auth_token: user.encoded(), info: userData })
+            res.json({ auth_token: user.encoded(), info: user.toJson() })
         } catch (e) {
-            res.status(400).json({ error: e })
-            return
-        }
-    }
-
-    static async getSession(req, res) {
-        try {
-            const { username } = req.body
-
-            let session = await UsersDAO.getUserSession(username)
-
-            if(!session) {
-                res.status(401).json({ error: "Session not found for this user." })
-                return
-            }
-            res.json({ session })
-        }
-        catch (e) {
-            res.status(400).json({ error: e })
-            return
-        }
-    }
-
-    static async logout(req, res) {
-        try {
-            const { username } = req.body
-
-            if(!username || typeof username !== "string") {
-                res.status(400).json({ error: "Bad username format, expected string." })
-                return
-            }
-            
-            let result = await UsersDAO.logoutUser(username)
-
-            if(!result.success) {
-                res.status(401).json({ error: "Session not found for this user." })
-                return
-            }
-            
-            res.json({ result })
-        }
-        catch (e) {
             res.status(400).json({ error: e })
             return
         }
@@ -154,21 +109,16 @@ class UsersController {
 
     static async populateTransactionForm(req, res) {
         try {
+            const bearer = req.headers['authorization']
             const { username } = req.body
 
-            if(!username || typeof username !== "string") {
-                res.status(400).json({ error: "Bad username format, expected string." })
-                return
+            if(await validateToken(bearer)) {
+                let result = await UsersDAO.populateTransactionForm(username) 
+                res.json({ result })
             }
-
-            let result = await UsersDAO.populateTransactionForm(username)
-            
-            if(!result) {
-                res.status(401).json({ error: "Could not populate transaction form." })
-                return
+            else {
+                res.status(403).json({ error: "Invalid token!"})
             }
-            
-            res.json({ result })
         }
         catch (e) {
             res.status(400).json({ error: e })
@@ -178,16 +128,16 @@ class UsersController {
 
     static async getAvailableAmount(req, res) {
         try {
+            const bearer = req.headers['authorization']
             let { username } = req.body
 
-            if(!username || typeof username !== "string") {
-                res.status(400).json({ error: "Bad username format, expected string." })
-                return
+            if(await validateToken(bearer)) {
+                let result = await UsersDAO.getAvailableAmount(username)
+                res.json({ result })   
             }
-
-            let result = await UsersDAO.getAvailableAmount(username)
-
-            res.json({ result })
+            else {
+                res.status(403).json({ error: "Invalid token!"})
+            }
         }
         catch (e) {
             res.status(400).json({ error: e })
@@ -197,13 +147,19 @@ class UsersController {
 
     static async getPortfolios(req, res) {
         try {
+            const bearer = req.headers['authorization']
             let { username } = req.body
 
-            let result = await UsersDAO.getPortfolios(username)
-
-            res.json({ result })
+            if(await validateToken(bearer)) {
+                let result = await UsersDAO.getPortfolios(username)
+                res.json({ result }) 
+            }
+            else {
+                res.status(403).json({ error: "Invalid token!"})
+            }
         }
         catch(e) {
+            console.log(e)
             res.status(400).json({ error: e })
             return
         }
@@ -211,11 +167,16 @@ class UsersController {
 
     static async getUserData(req, res) {
         try {
+            const bearer = req.headers['authorization']
             let { username } = req.body
 
-            let result = await UsersDAO.getUserData(username)
-
-            res.json({ result })
+            if(await validateToken(bearer)) {
+                let result = await UsersDAO.getUserData(username)
+                res.json({ result })
+            }
+            else {
+                res.status(403).json({ error: "Invalid token!"})
+            }
         }
         catch(e) {
             res.status(400).json({ error: e })
@@ -225,11 +186,16 @@ class UsersController {
 
     static async getUserCategories(req, res) {
         try {
+            const bearer = req.headers['authorization']
             let { username } = req.body
 
-            let result = await UsersDAO.getUserCategories(username)
-
-            res.json({ result })
+            if(await validateToken(bearer)) {
+                let result = await UsersDAO.getUserCategories(username)
+                res.json({ result })
+            }
+            else {
+                res.status(403).json({ error: "Invalid token!"})
+            }
         }
         catch(e) {
             res.status(400).json({ error: e })
@@ -239,11 +205,16 @@ class UsersController {
 
     static async getUserSources(req, res) {
         try {
+            const bearer = req.headers['authorization']
             let { username } = req.body
 
-            let result = await UsersDAO.getUserSources(username)
-
-            res.json({ result })
+            if(await validateToken(bearer)) {
+                let result = await UsersDAO.getUserSources(username)
+                res.json({ result }) 
+            }
+            else {
+                res.status(403).json({ error: "Invalid token!"})
+            }
         }
         catch(e) {
             res.status(400).json({ error: e })
@@ -253,10 +224,16 @@ class UsersController {
 
     static async addPortfolio(req, res) {
         try {
+            const bearer = req.headers['authorization']
             let { username, portfolio } = req.body
 
-            let result = await UsersDAO.addPortfolio(username, portfolio)
-            res.json({ result })
+            if(await validateToken(bearer)) {
+                let result = await UsersDAO.addPortfolio(username, portfolio)
+                res.json({ result })
+            }
+            else {
+                res.status(403).json({ error: "Invalid token!"})
+            }
         }
         catch(e) {
             res.status(400).json({ error: e })
@@ -266,11 +243,16 @@ class UsersController {
 
     static async changeFavourite(req, res) {
         try {
+            const bearer = req.headers['authorization']
             let { username, portfolio } = req.body
 
-            let result = await UsersDAO.changeFavourite(username, portfolio)
-
-            res.json({ result })
+            if(await validateToken(bearer)) {
+                let result = await UsersDAO.changeFavourite(username, portfolio)
+                res.json({ result })
+            }
+            else {
+                res.status(403).json({ error: "Invalid token!"})
+            }
         }
         catch(e) {
             res.status(400).json({ error: e })
@@ -286,11 +268,16 @@ class UsersController {
              *               from the deleted portfolio
              * transfer_amnt - The amount to be transfered from the deleted portfolio
              */
+            const bearer = req.headers['authorization']
             let { username, delete_id, transfer_id, transfer_amnt } = req.body
 
-            let result = await UsersDAO.deletePortfolio(username, delete_id, transfer_id, transfer_amnt)
-            
-            res.json({ result })
+            if(await validateToken(bearer)) {
+                let result = await UsersDAO.deletePortfolio(username, delete_id, transfer_id, transfer_amnt)
+                res.json({ result })
+            }
+            else {
+                res.status(403).json({ error: "Invalid token!"})
+            }
         }
         catch(e) {
             res.status(400).json({ error: e })
@@ -300,20 +287,26 @@ class UsersController {
 
     static async changePassword(req, res) {
         try {
+            const bearer = req.headers['authorization']
             let { username, old_pass, new_pass } = req.body
 
-            // Get current user password
-            let user = await UsersDAO.getUserPassword(username)
+            if(await validateToken(bearer)) {
+                // Get current user password
+                let user = await UsersDAO.getUserPassword(username)
 
-            if(await bcrypt.compare(old_pass, user[0].password)) {
-                // Hashing the password
-                let hashed = await hashPassword(new_pass)
-                let result = await UsersDAO.updatePassword(username, hashed)
-                res.json({ result })
+                if(await bcrypt.compare(old_pass, user[0].password)) {
+                    // Hashing the password
+                    let hashed = await hashPassword(new_pass)
+                    let result = await UsersDAO.updatePassword(username, hashed)
+                    res.json({ result })
+                }
+                else {
+                    res.status(401).json({ error: "The current password is not correct." })
+                    return
+                }
             }
             else {
-                res.status(401).json({ error: "The current password is not correct." })
-                return
+                res.status(403).json({ error: "Invalid token!"})
             }
         }
         catch(e) {
@@ -324,15 +317,16 @@ class UsersController {
 
     static async newCategory(req, res) {
         try {
+            const bearer = req.headers['authorization']
             let { username, category} = req.body
 
-            let result = await UsersDAO.newCategory(username, category)
-
-            if(result.error) {
-                res.status(401).json({error: result.error})
-                return
+            if(await validateToken(bearer)) {
+                let result = await UsersDAO.newCategory(username, category)
+                res.json({ result })
             }
-            res.json({ result })
+            else {
+                res.status(403).json({ error: "Invalid token!"})
+            }
         }
         catch(e) {
             res.status(400).json({ error: e })
@@ -342,18 +336,22 @@ class UsersController {
 
     static async newSource(req, res) {
         try {
+            const bearer = req.headers['authorization']
             let { username, source} = req.body
 
-            let result = await UsersDAO.newSource(username, source)
- 
-            if(result.error) {
-                res.status(401).json({error: result.error})
-                return
+            if(await validateToken(bearer)) {
+                let result = await UsersDAO.newSource(username, source)
+                res.json({ result })
             }
-            res.json({ result })
+            else {
+                res.status(403).json({ error: "Invalid token!"})
+            }
+            // if(result.error) {
+            //     res.status(401).json({error: result.error})
+            //     return
+            // }
         }
         catch(e) {
-            console.log(e)
             res.status(400).json({ error: e })
             return
         }
@@ -361,12 +359,18 @@ class UsersController {
 
     static async deleteCategory(req, res) {
         try {
+            const bearer = req.headers['authorization']
             let { username, category_id} = req.body
-            let result = await UsersDAO.deleteCategory(username, category_id)
-            res.json({ result })
+
+            if(await validateToken(bearer)) {
+                let result = await UsersDAO.deleteCategory(username, category_id)
+                res.json({ result })
+            }
+            else {
+                res.status(403).json({ error: "Invalid token!"})
+            } 
         }
         catch(e) {
-            console.log(e)
             res.status(400).json({ error: e })
             return
         }
@@ -374,12 +378,18 @@ class UsersController {
 
     static async deleteSource(req, res) {
         try {
+            const bearer = req.headers['authorization']
             let { username, source_id} = req.body
-            let result = await UsersDAO.deleteSource(username, source_id)
-            res.json({ result })
+
+            if(await validateToken(bearer)) {
+                let result = await UsersDAO.deleteSource(username, source_id)
+                res.json({ result })
+            }
+            else {
+                res.status(403).json({ error: "Invalid token!"})
+            }
         }
         catch(e) {
-            console.log(e)
             res.status(400).json({ error: e })
             return
         }
@@ -387,10 +397,16 @@ class UsersController {
 
     static async transfer(req, res) {
         try {
+            const bearer = req.headers['authorization']
             let { username, from, to, amount } = req.body
 
-            let result = await UsersDAO.transfer(username, from, to, amount)
-            res.json({ result })
+            if(await validateToken(bearer)) {
+                let result = await UsersDAO.transfer(username, from, to, amount)
+                res.json({ result })
+            }
+            else {
+                res.status(403).json({ error: "Invalid token!"})
+            }
         }
         catch(e) {
             res.status(400).json({ error: e })
